# packages/mtg-core/src/mtg_core/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MTG Core API",
    "description": "Magic: The Gathering card database and analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cards/search": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Search Cards",
        "description": "Search for cards with various filters.",
        "operationId": "search_cards_cards_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCardsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/details": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Get Card Details",
        "description": "Get detailed information about a specific card.",
        "operationId": "get_card_details_cards_details_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/rulings": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Get Card Rulings",
        "description": "Get official rulings for a card.",
        "operationId": "get_card_rulings_cards_rulings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRulingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/printings": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Get Card Printings",
        "description": "Get all printings of a card across all sets.",
        "operationId": "get_card_printings_cards_printings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardPrintingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "flavor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor Name"
          },
          "uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "supertypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supertypes"
          },
          "types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Types"
          },
          "subtypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subtypes"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "flavor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "set_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Code"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "layout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "power": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power"
          },
          "toughness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toughness"
          },
          "loyalty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loyalty"
          },
          "defense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defense"
          },
          "edhrec_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edhrec Rank"
          },
          "legalities": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legalities"
          },
          "rulings_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rulings Count"
          },
          "images": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageUrls"
              },
              {
                "type": "null"
              }
            ]
          },
          "prices": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Prices"
              },
              {
                "type": "null"
              }
            ]
          },
          "purchase_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PurchaseLinks"
              },
              {
                "type": "null"
              }
            ]
          },
          "related_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedLinks"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardDetail",
        "description": "Detailed card information."
      },
      "CardDetailsRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Card name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardDetailsRequest",
        "description": "Request for card details."
      },
...
      "CardPrintingsRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Card name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardPrintingsRequest",
        "description": "Request for card printings."
      },
      "CardRulingsRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Card name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardRulingsRequest",
        "description": "Request for card rulings."
      },
      "CardSummary": {
        "properties": {
          "uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "flavor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor Name"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "set_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Code"
          },
          "collector_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Number"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "power": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power"
          },
          "toughness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toughness"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "image_small": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Small"
          },
          "price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd"
          },
          "purchase_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Link"
          },
          "owned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owned"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CardSummary",
        "description": "Card summary for search results."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageUrls": {
        "properties": {
          "small": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Small"
          },
          "normal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normal"
          },
          "large": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Large"
          },
          "png": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Png"
          },
          "art_crop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art Crop"
          }
        },
        "type": "object",
        "title": "ImageUrls",
        "description": "Card image URLs in various sizes."
      },
...
      "Prices": {
        "properties": {
          "usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd"
          },
          "usd_foil": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Foil"
          },
          "eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eur"
          },
          "eur_foil": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eur Foil"
          }
        },
        "type": "object",
        "title": "Prices",
        "description": "Card prices."
      },
...
      "PrintingsResponse": {
        "properties": {
          "card_name": {
            "type": "string",
            "title": "Card Name"
          },
          "printings": {
            "items": {
              "$ref": "#/components/schemas/PrintingInfo"
            },
            "type": "array",
            "title": "Printings"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "card_name",
          "printings",
          "count"
        ],
        "title": "PrintingsResponse",
        "description": "All printings of a card."
      },
      "PurchaseLinks": {
        "properties": {
          "tcgplayer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tcgplayer"
          },
          "cardmarket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cardmarket"
          },
          "cardhoarder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cardhoarder"
          }
        },
        "type": "object",
        "title": "PurchaseLinks",
        "description": "Purchase links for a card."
      },
...
      "RelatedLinks": {
        "properties": {
          "edhrec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edhrec"
          },
          "gatherer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gatherer"
          }
        },
        "type": "object",
        "title": "RelatedLinks",
        "description": "Related links for a card."
      },
...
      "RulingsResponse": {
        "properties": {
          "card_name": {
            "type": "string",
            "title": "Card Name"
          },
          "rulings": {
            "items": {
              "$ref": "#/components/schemas/RulingEntry"
            },
            "type": "array",
            "title": "Rulings"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "card_name",
          "rulings",
          "count"
        ],
        "title": "RulingsResponse",
        "description": "Card rulings response."
      },
      "SearchCardsInput": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Card name (partial match)"
          },
          "colors": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "W",
                    "U",
                    "B",
                    "R",
                    "G",
                    "C"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colors",
            "description": "Filter by colors"
          },
          "color_identity": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "W",
                    "U",
                    "B",
                    "R",
                    "G",
                    "C"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Identity",
            "description": "Filter by color identity"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type",
            "description": "Card type"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype",
            "description": "Subtype"
          },
          "supertype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supertype",
            "description": "Supertype"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "common",
                  "uncommon",
                  "rare",
                  "mythic"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity",
            "description": "Card rarity"
          },
          "set_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Code",
            "description": "Set code"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc",
            "description": "Exact mana value"
          },
          "cmc_min": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc Min",
            "description": "Minimum mana value"
          },
          "cmc_max": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc Max",
            "description": "Maximum mana value"
          },
          "power": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power",
            "description": "Creature power"
          },
          "toughness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toughness",
            "description": "Creature toughness"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Search in card text"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords",
            "description": "Filter by keywords"
          },
          "format_legal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "standard",
                  "modern",
                  "legacy",
                  "vintage",
                  "commander",
                  "pioneer",
                  "pauper",
                  "historic",
                  "brawl",
                  "alchemy",
                  "explorer",
                  "timeless",
                  "oathbreaker",
                  "penny",
                  "duel"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Legal",
            "description": "Filter by format legality"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist",
            "description": "Filter by artist name"
          },
          "sort_by": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "name",
                  "cmc",
                  "color",
                  "rarity",
                  "type",
                  "random"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By",
            "description": "Sort by field (name, cmc, color, rarity, type)"
          },
          "sort_order": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Sort Order",
            "description": "Sort order (asc, desc)",
            "default": "asc"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Page number",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "Results per page",
            "default": 25
          },
          "random": {
            "type": "boolean",
            "title": "Random",
            "description": "Return random cards instead of sorted results",
            "default": false
          },
          "in_collection": {
            "type": "boolean",
            "title": "In Collection",
            "description": "Only return cards in user's collection",
            "default": false
          }
        },
        "type": "object",
        "title": "SearchCardsInput",
        "description": "Input parameters for search_cards tool."
      },
      "SearchResult": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "type": "array",
            "title": "Cards"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of cards on this page.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "cards",
          "page",
          "page_size",
          "count"
        ],
        "title": "SearchResult",
        "description": "Search results with pagination."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}